Allow leading whitespace in .sym65 comments
We allow empty lines and lines that begin with ';' in .sym65.  Lines
with nothing but whitespace, or comments with leading whitespace,
caused a warning.  It can be aesthetically nice to line up the start
of comments, and lines with pointless whitespace aren't problematic,
so we now allow these without complaint.

Added some samples to the 20170-external-symbols data file.  No
change to the test output.