Fix keyboard nav after selection jump
If you change the selection by double-clicking in one of the side
windows (References, Notes, Symbols, Messages) and then attempt
to navigate with the arrow keys, the program will appear to hang
briefly, then jump to the start or end of the project and shift the
window focus to something else (like the Help menu in the menu bar).

For reasons I don't fully understand, the behavior is fixed by
removing unnecessary calls to codeListView.Focus().  The calls were
intended to shift input focus to the main ListView, but they're no
longer necessary, and appear to upset WPF.

(issue #113)