Migrate to Visual Studio Community 2022
Minor change to solution.  Removed CodeLab project stub, which was
still targeting .NET Framework 4.6.1.

The other projects with WPF code target 4.6.2, and I haven't updated
them to 4.8 out of a fear that it might make 6502bench harder to
install on older systems.  The down side is that it requires an
extra "dev pack" download to build the source code.  4.8 is the end
of the line for .NET Framework, so at some point we should make that
the official target.