Make notes panel text column fixed-width
The default width doesn't line up with the right edge of the panel,
so either you're not seeing as much text as you could, or it's
scrollable and will lurch over if you double-click in the text area.

Now the width is set to '*', which makes the column wide enough to
fill the panel, and automatically resizes as things shift.