Expand goto dialog address resolution
The "goto" dialog (Ctrl+G) jumps to an offset, address, or label.  If
an address is specified, it is resolved relative to the first line of
the current selection, so that if the same address appears multiple
times we select the "best" one.

This turned out to be too restrictive, because if the region was marked
with "disallow outbound address resolution" we wouldn't look anywhere
except the current region.  While arguably correct, it was annoying
to have to move the cursor before being able to jump to a unique
address elsewhere in the file.

Now, if the address lookup fails to identify an offset, we repeat it
with "disallow outbound" flags ignored.