Update API target to .NET Framework 4.8.1
The project was targeting 4.6.2, which was the current version when
work first began.  This update should not cause any change in behavior.
The only real value in doing this is that it means people working on
the project won't have to install the older SDK components.

This may require performing a manual "clobber" in existing source trees:
close Visual Studio, then in each of the seven projects, manually
remove the "bin" and "obj" directories.  Using the VS "clean" feature
doesn't seem to clear out all of the dependencies, and you get weird
build complaints about missing System classes.

I'm not anticipating any compatibility issues with this switch.
Framework 4.8 shipped in April 2019, and the final version of .NET
Framework was released August 2022, so anybody who has Framework
installed should have a compatible version.

This change does not move the libraries from .NET Standard 2.0 to 2.1,
because 2.0 was the last version supported by Framework.

(At some point it might be useful to upgrade to the current .NET, but
that is a more significant change.)