Stop showing irrelevant cycle mods
We were showing (relevant) [irrelevant] cycle mods in the Notes panel,
but unless you're working with a 65816 you probably don't care about
things affected by the M/X flags and D register.

Also, made the rightmost column of the Notes and Symbols panels
auto-sized to fill the window.  We have to do this in the code because
we're constantly saving/restoring the column widths so that we can
record the values in the app setttings file.