Add anchors for long comments
When a chunk of code is preceded by a block comment, it's nice to
provide an HTML link that goes to the start of the comment rather
than the start of the code.  Browsers like Chrome scroll the page
to position the anchor line at the very top, so it's not immediately
obvious that a comment is present.

We now generate anchors for the first line of long comments that
come before a code or data item with a label.

(issue #181)